academic-backend-microservice/academicmanagement/requirements.txt
asgiref==3.8.1
attrs==24.1.0
auto-cpufreq==1
autocommand==2.2.2
blessed==1.20.0
btrfsutil==6.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.3
distlib==0.3.9
distro==1.9.0
Django==5.2.2
django-cors-headers==4.7.0
django-model-utils==5.0.0
djangorestframework==3.16.0
filelock==3.18.0
gpustat==1.1.1
idna==3.10
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
lxml==5.4.0
Mako==1.3.10.dev0
Markdown==3.8
MarkupSafe==3.0.2
meson==1.8.1
more-itertools==10.7.0
mysql-connector-python==9.2.0
nftables==0.1
nvidia-ml-py==12.575.51
packaging==25.0
platformdirs==4.3.6
psutil==7.0.0
pycparser==2.22
Pygments==2.19.1
PyGObject==3.52.3
pyinotify==0.9.6
python-dotenv==1.1.0
requests==2.32.3
setuptools==80.9.0
setuptools-scm==8.3.1
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
urllib3==2.4.0
virtualenv==20.28.0
wcwidth==0.2.13
wheel==0.45.1
